backend/requirements.txt

fastapi
uvicorn
python-multipart     
ultralytics           
google-genai          
python-dotenv         
pillow                